v2/omni-scraper/requirements.txt
# ═══════════════════════════════════════════════════
#  OMNI-SCRAPER — Python Dependencies
#  Compatible with Python 3.11+ / GitHub Codespaces
# ═══════════════════════════════════════════════════

# ── Browser Automation ──────────────────────────────
playwright==1.44.0
playwright-stealth==1.0.6

# ── Async & Concurrency ─────────────────────────────
asyncio==3.4.3
anyio==4.4.0
aiohttp==3.9.5
httpx==0.27.0

# ── LLM Integrations ────────────────────────────────
openai==1.35.0              # OpenAI GPT-4o
anthropic==0.29.0           # Claude 3.5 Sonnet
langchain==0.2.6
langchain-openai==0.1.13
langchain-anthropic==0.1.19
langgraph==0.1.14
langchain-community==0.2.6

# ── NLP / Semantic Understanding ────────────────────
spacy==3.7.5
trafilatura==1.10.0
lxml==5.2.2
beautifulsoup4==4.12.3
html2text==2024.2.26

# ── Vector Memory (Qdrant) ──────────────────────────
qdrant-client==1.9.1
sentence-transformers==3.0.1

# ── Task Queue / Distributed ────────────────────────
celery==5.4.0
redis==5.0.6
flower==2.0.1               # Celery monitoring UI

# ── Data & Storage ──────────────────────────────────
sqlalchemy==2.0.31
alembic==1.13.1
pandas==2.2.2
numpy==1.26.4
pydantic==2.7.4
pydantic-settings==2.3.3

# ── API Server ──────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1
websockets==12.0
python-multipart==0.0.9

# ── Proxy Management ────────────────────────────────
aiohttp-socks==0.8.4
python-socks==2.4.3

# ── Telemetry & Observability ───────────────────────
prometheus-client==0.20.0
opentelemetry-sdk==1.25.0
opentelemetry-api==1.25.0
structlog==24.2.0

# ── Utilities ───────────────────────────────────────
python-dotenv==1.0.1
pyyaml==6.0.1
click==8.1.7
rich==13.7.1
httpx[socks]==0.27.0
tenacity==8.3.0
cryptography==42.0.8

# ── Export ──────────────────────────────────────────
openpyxl==3.1.3
tabulate==0.9.0
